requirements.txt
requests
lxml
rpmfile @ git+https://github.com/boiko/rpmfile@master
python-bugzilla
